+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0
 
 * Adopt [Semantic Versioning](http://semver.org/) scheme.
+* Remove @deprecated options.
 
 ## 1.7.5